feat(DEV-1674): fix page styles
