+++ apps/nextjs-website/src/helpers/breadcrumbs.helpers.ts
@@ -42,21 +42,3 @@ export function pageToBreadcrumbs(
     ...(paths || []),
   ];
 }
-
-export function solutionPageToBreadcrumbs(
-  paths?: readonly Path[]
-): readonly BreadCrumbSegment[] {
-  return [
-    {
-      name: 'home',
-      path: '/',
-      translate: true,
-    },
-    {
-      name: 'solutions',
-      path: '/solutions',
-      translate: true,
-    },
-    ...(paths || []),
-  ];
-}
